prefer Popen over subprocess.run

+++ reboot.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/env bash
 
-sleep 6
-[ -z $1 ] && exit 1
-kill -9 $1
 sleep 5
+[ ! -z $1 ] && kill -9 $1
+sleep 8
 python3 -m pyUltroid
